refactor: reorganize internal/auth by responsibility
Split the 760-line auth.go into focused files with no API change:
- token_store.go: token cache persistence (save/load/atomic write,
  persisting token source)
- login_flow.go: interactive OAuth login and local callback server
- payload.go: .clasprc.json payload parsing
- auth.go: OAuth config resolution and public client entry points

Deduplicate saveToken/saveTokenExclusive via shared ensureTokenCacheDir
and tokenFileJSON helpers, and extract clientFromCachedToken and
startCallbackServer out of the 150-line loginWithCachePath. All output
messages and error texts are unchanged. The ldflags-injected variables
(ldflagsClientID/ldflagsClientSecret) keep their name and package path;
verified with make build.
